@@ -19,9 +19,13 @@
 		"prepublish": "yarn run build && yarn run copy_publish_files"
 	},
 	"peerDependencies": {
-		"@material-ui/core": "^4.9.4",
-		"react": "^16.8.6 || ^17.0.0",
-		"react-dom": "^16.8.6 || ^17.0.0"
+		"@emotion/react": "^11.8.2",
+    	"@emotion/styled": "^11.8.1",
+		"@mui/icons-material": "^5.6.0",
+		"@mui/material": "^5.5.2",
+		"@mui/styles": "^5.6.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2"
 	},
 	"devDependencies": {
 		"@babel/core": "^7.0.0",
@@ -44,10 +48,15 @@
 		"@babel/plugin-syntax-import-meta": "^7.0.0",
 		"@babel/preset-env": "^7.0.0",
 		"@babel/preset-react": "^7.0.0",
+		"@emotion/react": "^11.8.2",
+    	"@emotion/styled": "^11.8.1",
+		"@mui/icons-material": "^5.6.0",
+		"@mui/material": "^5.5.2",
+		"@mui/styles": "^5.6.0",
 		"@types/classnames": "^2.2.9",
 		"@types/jest": "^25.1.2",
 		"@types/lodash.isequal": "^4.5.5",
-		"@types/react": "^16.9.19",
+		"@types/react": "^17.0.2",
 		"@typescript-eslint/eslint-plugin": "^2.19.0",
 		"@typescript-eslint/parser": "^2.19.0",
 		"babel-eslint": "^10.0.1",
@@ -58,8 +67,8 @@
 		"eslint-plugin-jsx-a11y": "^6.2.3",
 		"eslint-plugin-react": "^7.18.3",
 		"eslint-plugin-react-hooks": "^1.7.0",
-		"react": "16.8.6 || ^17.0.0",
-		"react-dom": "16.8.6 || ^17.0.0",
+		"react": "^17.0.2",
+		"react-dom": "^17.0.2",
 		"react-scripts": "^3.0.1",
 		"rollup": "^1.1.2",
 		"rollup-plugin-babel": "^4.3.2",
