@@ -1,5 +1,5 @@
 import React from 'react';
-import { List, ListItem, ListItemText } from '@material-ui/core';
+import { List, ListItem, ListItemText } from '@mui/material';
 import { isSameDay } from 'date-fns';
 
 // eslint-disable-next-line no-unused-vars
